Rework test structure inside storage package.
Create storage_test.go file to store testMain function and docker interaction logic.
All logic that is common for user and tournament tests inside storage package is now holded here.
